requirements.txt
Pillow==7.1.1
python-telegram-bot==12.3.0
numpy==1.18.3